refactor: updated product model so price is integer; updated view so that shows dollar figure

+++ server/db/models/product.js
@@ -11,7 +11,7 @@ const Product = db.define('product', {
     }
   },
   price: {
-    type: Sequelize.DECIMAL(5, 2),
+    type: Sequelize.INTEGER,
     validate: {
       min: 0
     }
